chore: move unexported components to storybook (#2013)
- Grid and Section are only used within the design system
- move them out of the regular components directory to avoid analytics scans
- make sure no other usages change as a result
