+++ .storybook/components/DesignTokens/Tier1/FontFamilies.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
-import Grid from '../../../../src/components/Grid';
-import Section from '../../../../src/components/Section';
 import filterTokens from '../../../util/filterTokens';
+import Grid from '../../Grid';
+import Section from '../../Section';
 import { TokenSpecimen } from '../../TokenSpecimen/TokenSpecimen';
 
 export const FontFamilies = () => (
