+++ .storybook/components/DesignTokens/Tier1/TypographyPresets.jsx
@@ -1,10 +1,10 @@
 import { at, forEach } from 'lodash';
 
 import React, { Component } from 'react';
-import Grid from '../../../../src/components/Grid';
-import Section from '../../../../src/components/Section';
 import presets from '../../../../src/design-tokens/tier-1-definitions/typography.json';
 import flatten from '../../../util/flattenToken';
+import Grid from '../../Grid';
+import Section from '../../Section';
 import { TokenSpecimen } from '../../TokenSpecimen/TokenSpecimen';
 
 export class Tier1TypographyPresets extends Component {
