+++ .storybook/components/DesignTokens/Tier2/Forms.stories.tsx
@@ -1,8 +1,8 @@
 import type { StoryObj } from '@storybook/react';
 import React from 'react';
-import Grid from '../../../../src/components/Grid';
-import Section from '../../../../src/components/Section';
 import filterTokens from '../../../util/filterTokens';
+import Grid from '../../Grid';
+import Section from '../../Section';
 import { TokenSpecimen } from '../../TokenSpecimen/TokenSpecimen';
 
 export default {
